adding logout user function in user constroller

+++ Backend/src/utils/cloudinary.js
@@ -20,7 +20,10 @@ const uploadOnCloudinary = async (localFilePath) => {
 
     // file has been uploaded successfully
     console.log("File is uploaded on cloudinary", response.url);
-    fs.unlinkSync(localFilePath);
+    // delete the local file if it exists
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
     return response;
   } catch (error) {
     fs.unlinkSync(localFilePath); // remove the localy saved temporary file as the upload operation got faield
